refactor: move general app styles to styles directory
