+++ src/styles/App.scss
@@ -10,15 +10,14 @@ $carbon--theme: $carbon--theme--g10;
 @include carbon--theme();
 
 // import here so they inherit the theme g10
-@import "./content/Dashboard/dashboard.scss";
-@import "./content/Metrics/metrics.scss";
-@import "./content/SMS/sms.scss";
-@import "./content/Modem/modems.scss";
-@import "./content/Profile/profile.scss";
-@import "./content/Settings/settings.scss";
-
-@import "./components/DashHeader/dash_header.scss";
-@import "./components/Card/card.scss";
+@import "../content/Dashboard/dashboard.scss";
+@import "../content/Metrics/metrics.scss";
+@import "../content/SMS/sms.scss";
+@import "../content/Modem/modems.scss";
+@import "../content/Profile/profile.scss";
+@import "../content/Settings/settings.scss";
+@import "../components/DashHeader/dash_header.scss";
+@import "../components/Card/card.scss";
 
 //general settings go here
 .bx--side-nav__navigation {
